% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_apply_functions.R
\name{applyAlpha}
\alias{applyAlpha}
\title{the basic functions for calculating and bootstrapping the internal consistency estimates}
\usage{
applyAlpha(M)
}
\description{
the basic functions for calculating and bootstrapping the internal consistency estimates
}
