\name{asrm}
\alias{asrm}
\docType{data}
\title{
5-Item questionnaire data from Nicolai (2018)
}
\description{
A dataset consisting of 78 participants who filled out the 5-item Altman Self-Rating Mania Scale,
rating from 1 to 5 on a Likert scale}
\usage{asrm}
\format{
  The format is a 5-column datamatrix containing 78 observations
}
\source{
article}
\references{
Nicolai, J., & Moshagen, M. (2018). Pathological buying symptoms are associated with distortions in judging elapsed time. Journal of Behavioral Addictions, 7(3), 752-759.
}

\keyword{datasets}
