\name{asrm_mis}
\alias{asrm_mis}
\docType{data}
\title{
  5-Item questionnaire data from Nicolai (2018) with 10 \% missings
}
\description{
  A dataset consisting of 78 participants who filled out the 5-item Altman Self-Rating Mania Scale,
  rating from 1 to 5 on a Likert scale, 10 \% missings were inserted at random}
\usage{asrm_mis}
\format{
  The format is a 5-column data matrix containing 78 observations, missing are NA
}
\source{
  article}
\references{
  Nicolai, J., & Moshagen, M. (2018). Pathological buying symptoms are associated with distortions in judging elapsed time. Journal of Behavioral Addictions, 7(3), 752-759.
}

\keyword{datasets}
