% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_output.R
\name{bayesFun}
\alias{bayesFun}
\title{functions to return the jags sampled estimates of internal concsistency
together with credible intervals as well as the sampled posterior distribution}
\usage{
bayesFun(data, n.iter, n.burnin, estimates, interval, omega.cov.samp,
  return.cov.samples)
}
\description{
functions to return the jags sampled estimates of internal concsistency
together with credible intervals as well as the sampled posterior distribution
}
