% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFrel.R
\name{bfrel}
\alias{bfrel}
\title{gives the odds ratio of an estimate being bigger than low.bound between the prior and the posterior
needs a posterior and a prior sample of the estimate as input, and the critical value of the hypothesis}
\usage{
bfrel(xprior, xpost, low.bound)
}
\description{
gives the odds ratio of an estimate being bigger than low.bound between the prior and the posterior
needs a posterior and a prior sample of the estimate as input, and the critical value of the hypothesis
}
