% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_apply_functions.R
\name{bootAlpha}
\alias{bootAlpha}
\title{the basic functions for calculating and bootstrapping the internal consistency estimates}
\usage{
bootAlpha(data, indices)
}
\description{
the basic functions for calculating and bootstrapping the internal consistency estimates
}
