% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_reliability.R
\name{bre}
\alias{bre}
\title{function to calculate all the internal consitency estimates
--------------------------- IMPORTANT ----------------------:
decide on how to calculate the confidence interval for omega, bootstrapping or analytical by lavaan,
this is also important for the package, whose print methods depend on that, unfortunately}
\usage{
bre(raw.data, boot.n = 200, interval = 0.95, boot.interval.type = "basic",
  jags = FALSE, n.iter = 2000, n.burnin = 50, freq = TRUE,
  estimates = c("alpha", "l2", "l6", "glb", "omega"), supr.warnings = TRUE,
  omega.freq.method = "pa", omega.conf.type = "alg")
}
\description{
maybe use a pca instead of a cfa. especially with the bad performance with small samples
}
