% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_reliability.R
\name{brel}
\alias{brel}
\title{function to calculate all the internal consitency estimates}
\usage{
brel(raw.data, boot.n = 200, interval = 0.95, jags = FALSE,
  n.iter = 2000, n.burnin = 50, estimates = c("alpha", "lambda2",
  "lambda6", "glb", "omega"), supr.warnings = TRUE,
  omega.freq.method = "pa", omega.conf.int.type = "boot",
  omega.cov.samp = FALSE, return.cov.samples = FALSE,
  prior.samp = FALSE, item.dropped = FALSE,
  alpha.int.analytic = FALSE, bayes = TRUE, freq = TRUE,
  boot.interval.type = "basic")
}
\description{
function to calculate all the internal consitency estimates
}
