% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{brelPlot}
\alias{brelPlot}
\title{plot function for a posterior sample, gives posterior and prior distribution and pie plots
input is the main reliability estimation object}
\usage{
brelPlot(res, top.align = FALSE, name = NULL, blackwhite = FALSE,
  criteria = TRUE)
}
\description{
plot function for a posterior sample, gives posterior and prior distribution and pie plots
input is the main reliability estimation object
}
