\name{cavalini}
\alias{cavalini}
\docType{data}
\title{
8-Item Questionnaire Data from Cavalini (1992)
}
\description{
A dataset consisting of eight item questionnaire data. It's Likert scaled from 0-3.
It is data measuring how annoyed people were by malodors}
\usage{cavalini}
\format{
  The format is a 8-column datamatrix containing 828 observations
}
\source{
Doctoral Dissertation}
\references{
Cavalini, P. M. (1992). It's an ill wind that brings no good: Studies on odour annoyance and the dispersion of odorant concentrations from industries. Rijksuniversiteit Groningen.
}

\keyword{datasets}
