% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_gibbs_sampler.R
\name{covSamp2}
\alias{covSamp2}
\title{this function uses gibbs sampling to estimate the posterior distribution
of a sample's covariance matrix}
\usage{
covSamp2(data, n.iter, n.burnin)
}
\description{
this function uses gibbs sampling to estimate the posterior distribution
of a sample's covariance matrix
}
