% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glb_algebraic2.R
\name{csdp2}
\alias{csdp2}
\title{code from the Rcsdp package:
Hector Corrada Bravo (2016). Rcsdp: R Interface to the CSDP Semidefinite Programming Library.
R package version 0.1.55. https://CRAN.R-project.org/package=Rcsdp}
\usage{
csdp2(C, A, b, K, control = csdp.control(printlevel = 0))
}
\description{
code from the Rcsdp package:
Hector Corrada Bravo (2016). Rcsdp: R Interface to the CSDP Semidefinite Programming Library.
R package version 0.1.55. https://CRAN.R-project.org/package=Rcsdp
}
