% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_omega.R
\name{fit.omega}
\alias{fit.omega}
\title{graphical posterior predictive check for the 1-factor omega model, based on covariance matrix eigenvalues}
\usage{
fit.omega(x)
}
\arguments{
\item{x}{A strel output object (list)}
}
\description{
gives posterior predictive check for the 1-factor model:
comparison between model implied covariance matrix and sample covariance matrix
also displays frequentist fit indices
}
\examples{
fit.omega(strel(asrm, "omega"))

}
