% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_output.R
\name{freqFun}
\alias{freqFun}
\title{this function calls on other functions in order to return the fuentist estimates
and bootstrapped cinfidence intervals, which can be of different types}
\usage{
freqFun(data, boot.n, boot.interval.type, estimates, interval,
  omega.freq.method, omega.conf.int.type)
}
\description{
this function calls on other functions in order to return the fuentist estimates
and bootstrapped cinfidence intervals, which can be of different types
}
