% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_output2.R
\name{freqFun2}
\alias{freqFun2}
\title{this function calls on other functions in order to return the fuentist estimates
and bootstrapped confidence intervals, which can be of different types}
\usage{
freqFun2(data, boot.n, estimates, interval, omega.freq.method, omega.conf.type)
}
\description{
this function calls on other functions in order to return the fuentist estimates
and bootstrapped confidence intervals, which can be of different types
}
