% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_nonparametric.R
\name{freqFun_nonpara}
\alias{freqFun_nonpara}
\title{this function calls on other functions in order to return the frequentist estimates
and non-parametric bootstrapped confidence intervals, now calculated with SEs and z-values}
\usage{
freqFun_nonpara(data, boot.n, estimates, interval, omega.freq.method,
  item.dropped, alpha.int.analytic)
}
\description{
this function calls on other functions in order to return the frequentist estimates
and non-parametric bootstrapped confidence intervals, now calculated with SEs and z-values
}
