% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_parametric.R
\name{freqFun_para}
\alias{freqFun_para}
\title{this function calls on other functions in order to return the frequentist estimates
and parametric bootstrapped confidence intervals, sampling from a multivariate normal distribution}
\usage{
freqFun_para(data, boot.n, estimates, interval, omega.freq.method,
  item.dropped, alpha.int.analytic)
}
\description{
this function calls on other functions in order to return the frequentist estimates
and parametric bootstrapped confidence intervals, sampling from a multivariate normal distribution
}
