% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_output.R
\name{gibbsFun}
\alias{gibbsFun}
\title{this function calls on other functions in order to return the gibbs sampled estimates
and the credible intervals together with the posterior distribution objects
to be passed on for forther analysis}
\usage{
gibbsFun(data, n.iter, n.burnin, estimates, interval)
}
\description{
this function calls on other functions in order to return the gibbs sampled estimates
and the credible intervals together with the posterior distribution objects
to be passed on for forther analysis
}
