% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glb_algebraic2.R
\name{glb.algebraic2}
\alias{glb.algebraic2}
\title{code from psych Package:
Revelle, W. (2018) psych: Procedures for Personality and Psychological Research,
Northwestern University, Evanston, Illinois, USA, https://CRAN.R-project.org/package=psych Version = 1.8.4.}
\usage{
glb.algebraic2(Cov, LoBounds = NULL, UpBounds = NULL)
}
\description{
code from psych Package:
Revelle, W. (2018) psych: Procedures for Personality and Psychological Research,
Northwestern University, Evanston, Illinois, USA, https://CRAN.R-project.org/package=psych Version = 1.8.4.
}
