% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadamard.R
\docType{data}
\name{hadamard.sizes2}
\alias{hadamard.sizes2}
\title{hadamard support functions functions}
\format{An object of class \code{numeric} of length 4.}
\usage{
hadamard.sizes2
}
\description{
hadamard support functions functions
}
\keyword{datasets}
