% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_output.R
\name{jagsFun}
\alias{jagsFun}
\title{functions to return the jags sampled estimates of internal concsistency
together with credible intervals as well as the sampled posterior distribution}
\usage{
jagsFun(data, n.iter, n.burnin, estimates, interval, omega.cov)
}
\description{
functions to return the jags sampled estimates of internal concsistency
together with credible intervals as well as the sampled posterior distribution
}
