% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_half.R
\name{MaxSplitExhaustive}
\alias{MaxSplitExhaustive}
\title{this function uses an exhaustive algorithm to determine the best possible splithalf reliability}
\usage{
MaxSplitExhaustive(M)
}
\description{
this function uses an exhaustive algorithm to determine the best possible splithalf reliability
}
