% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_functions.R
\name{modelfile}
\alias{modelfile}
\title{functions for sampling using JAGS
separate sampling for the covariance matrix and the cfa model}
\usage{
modelfile()
}
\description{
functions for sampling using JAGS
separate sampling for the covariance matrix and the cfa model
}
