% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funs.R
\name{mvrnorm2}
\alias{mvrnorm2}
\title{create multivariate sample with mu and sigma given
source: MASS package
Venables, W. N. & Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth Edition. Springer, New York. ISBN 0-387-95457-0}
\usage{
mvrnorm2(n = 1, mu, Sigma, tol = 1e-06, empirical = TRUE)
}
\description{
create multivariate sample with mu and sigma given
source: MASS package
Venables, W. N. & Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth Edition. Springer, New York. ISBN 0-387-95457-0
}
