% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_fit.R
\name{omega_fit}
\alias{omega_fit}
\title{graphical posterior predictive check for the 1-factor omega model, based on covariance matrix eigenvalues}
\usage{
omega_fit(x)
}
\arguments{
\item{x}{A strel output object (list)}
}
\description{
gives posterior predictive check for the 1-factor model:
comparison between model implied covariance matrix and sample covariance matrix
also displays frequentist fit indices
}
\examples{
omega_fit(strel(asrm, "omega", n.chains = 2, n.iter = 100))

}
