% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omegaFit.R
\name{omegaFit}
\alias{omegaFit}
\title{graphical posterior predictive check for the 1-factor omega model,
based on covariance matrix eigenvalues}
\usage{
omegaFit(x, data, ppc = TRUE, cutoff = 0.08, ci = 0.9)
}
\arguments{
\item{x}{A strel output object (list)}

\item{data}{A matrix or data.frame containing the data set that produced x}

\item{ppc}{A logical indicating if the PPC should be printed or not,
the default is TRUE}

\item{cutoff}{A value to compare the posterior sample of RMSEAs against.
The result will contain the probability that the RMSEA is smaller than the
cutoff value}

\item{ci}{A value between 0 and 1 indicating the credible interval for the RMSEA}
}
\description{
gives posterior predictive check for the 1-factor model:
comparison between model implied covariance matrix and sample covariance matrix
also displays frequentist fit indices
}
\examples{
omegaFit(strel(asrm, "omega", n.chains = 2, n.iter = 200), data = asrm)

}
\references{
{
\insertRef{Garnier-Villarreal2020AdaptingFitIndices}{Bayesrel}
}
}
