% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_freq.R
\name{omegaFreq}
\alias{omegaFreq}
\title{gives freq omega, and loadings and errors}
\usage{
omegaFreq(data, estimator = "mlr", se = "default", missing = "ml",
  equal.loading = FALSE, equal.error = FALSE)
}
\description{
gives freq omega, and loadings and errors
}
