% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{p_omegas-deprecated}
\alias{p_omegas-deprecated}
\title{prior and posterior probability of omega_t and omega_h being bigger than thresholds}
\usage{
p_omegas(x, cutoff.t = .80, cutoff.h = .60)
}
\arguments{
\item{x}{A strel output object (list)}

\item{cutoff.t}{A number indicating the threshold for omega_t}

\item{cutoff.h}{A number indicating the threshold for omega_h}
}
\description{
takes mcmc posterior samples of omega_t and omega_h
and calculates the prior and posterior probability of the estimate being bigger
or smaller than an arbitrary value
}
\seealso{
\code{\link{Bayesrel-deprecated}}
}
\keyword{internal}
