% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotic}
\alias{plotic}
\title{plot function for a posterior sample, gives posterior and prior distribution and pie plots
input is the main reliability estimation object and the estimate to be plotted}
\usage{
plotic(x, estimate, blackwhite = FALSE, criteria = TRUE,
  cuts = c(0.7, 0.8), twopie = FALSE)
}
\description{
plot function for a posterior sample, gives posterior and prior distribution and pie plots
input is the main reliability estimation object and the estimate to be plotted
}
