% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotic_ifitem_all}
\alias{plotic_ifitem_all}
\title{plots posterior distributions of chosen estimate and the item-dropped cases in one plot}
\usage{
plotic_ifitem_all(x, estimate, ordering = FALSE)
}
\description{
plots posterior distributions of chosen estimate and the item-dropped cases in one plot
}
