% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plotIfItem}
\alias{plotIfItem}
\title{plotting function of the posterior for the if-item-dropped statistics}
\usage{
plotIfItem(res, estimate, item.pos, criteria = TRUE, blackwhite = FALSE,
  top.align = FALSE, greek = FALSE)
}
\description{
plotting function of the posterior for the if-item-dropped statistics
}
