% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plotrel}
\alias{plotrel}
\title{plot function for a posterior sample, gives posterior and prior distribution and pie plots
input is the main reliability estimation object and the estimate to be plotted}
\usage{
plotrel(x, estimate, top.align = FALSE, greek = FALSE,
  blackwhite = FALSE, criteria = TRUE, cuts = c(0.7, 0.8))
}
\description{
plot function for a posterior sample, gives posterior and prior distribution and pie plots
input is the main reliability estimation object and the estimate to be plotted
}
