% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotstrel.id}
\alias{plotstrel.id}
\title{plots posterior distributions of chosen estimate and the item-dropped cases in one plot}
\usage{
plotstrel.id(x, estimate, ordering = FALSE)
}
\arguments{
\item{x}{A strel output object (list)}

\item{estimate}{A character string indicating what estimate to plot from the strel output object}

\item{ordering}{A logical indicating if the densities in the plot should be ordered}
}
\description{
gives posterior densities of original dataset together with the the posteriors of datasets with
items deleted. Can be ordered for the change item deleting brings about
}
\examples{
plotstrel.id(strel(cavalini, "lambda2", item.dropped = TRUE), "lambda2")

}
