% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotstrel}
\alias{plotstrel}
\title{plot function for an single test reliability estimate's posterior sample}
\usage{
plotstrel(x, estimate, blackwhite = FALSE, criteria = TRUE,
  cuts = c(0.7, 0.8), twopie = FALSE)
}
\arguments{
\item{x}{A strel output object (list)}

\item{estimate}{A character string indicating what estimate to plot from the strel output object}

\item{blackwhite}{A logical indicating if the plot should be in black and white}

\item{criteria}{A logical indicating if cutoff criteria should be drawn}

\item{cuts}{A two element vector indicating what the cutoffs should be}

\item{twopie}{A logical indicating if an additional pie plot with the prior should be drawn}
}
\description{
gives posterior and prior distribution and pie plots
input is the main reliability estimation object and the estimate to be plotted
}
\examples{
plotstrel(strel(cavalini, "lambda2"), "lambda2")

}
