% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProb.R
\name{pOmegas}
\alias{pOmegas}
\title{prior and posterior probability of omega_t and omega_h being larger than thresholds}
\usage{
pOmegas(x, cutoff.t = 0.8, cutoff.h = 0.6)
}
\arguments{
\item{x}{A bomegas output object (list)}

\item{cutoff.t}{A number indicating the threshold for omega_t}

\item{cutoff.h}{A number indicating the threshold for omega_h}
}
\description{
takes mcmc posterior samples of omega_t and omega_h
and calculates the prior and posterior probability of the estimate being bigger
or smaller than an arbitrary value
}
\examples{
pOmegas(bomegas(upps, n.factors = 5, n.chains = 2, n.iter = 150, n.burnin = 50,
disableMcmcCheck = TRUE, missing = "listwise"))
}
