% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ppcOmega}
\alias{ppcOmega}
\title{graphical posterior predictive check for the 1-factor omega model, based on covariance matrix eigenvalues}
\usage{
ppcOmega(data, lambda, psi)
}
\description{
graphical posterior predictive check for the 1-factor omega model, based on covariance matrix eigenvalues
}
