% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/princ_fac.R
\name{princFac}
\alias{princFac}
\title{function to do a principal factor analysis, also principal axis method
http://www.aaronschlegel.com/iterated-principal-factor-method-factor-analysis-r/}
\usage{
princFac(m, max.iter = 50)
}
\description{
function to do a principal factor analysis, also principal axis method
http://www.aaronschlegel.com/iterated-principal-factor-method-factor-analysis-r/
}
