% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_IC.R
\name{probic}
\alias{probic}
\title{takes a mcmc posterior sample of any of the internal consistency estimates
and calculates any given probability of the estimate being bigger
or smaller than an arbitrary value}
\usage{
probic(x, estimate, low.bound)
}
\description{
takes a mcmc posterior sample of any of the internal consistency estimates
and calculates any given probability of the estimate being bigger
or smaller than an arbitrary value
}
