% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_strel.R
\name{pstrel}
\alias{pstrel}
\title{probability of estimate being bigger than threshold}
\usage{
pstrel(x, estimate, low.bound)
}
\arguments{
\item{x}{A strel output object (list)}

\item{estimate}{A character string indicating what estimate to plot from the strel output object}

\item{low.bound}{A number for the threshold to be tested against}
}
\description{
takes a mcmc posterior sample of any of the single test reliability estimates
and calculates any given probability of the estimate being bigger
or smaller than an arbitrary value
}
\examples{
pstrel(strel(cavalini, "lambda2"), "lambda2", .80)
}
