% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_split_half.R
\name{quant.lambda4}
\alias{quant.lambda4}
\title{finds split reliability coefficients, produces sample of lambda4s,
default return is max lambda4, but other quantiles can also be extracted
ref.: Hunt & Bentler 2015}
\usage{
quant.lambda4(x, starts = 1000, quantile = 1)
}
\description{
finds split reliability coefficients, produces sample of lambda4s,
default return is max lambda4, but other quantiles can also be extracted
ref.: Hunt & Bentler 2015
}
