% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_reliability.R
\name{rel}
\alias{rel}
\title{function to calculate all the internal consitency estimates
takes as input both datasets, either matrix or frame, and covariance matrices
so far input of a covariance matrix is not supported}
\usage{
rel(x, boot.n = 200, interval = 0.95, n.iter = 2000, n.burnin = 50,
  estimates = c("alpha", "lambda2", "lambda6", "glb", "omega"),
  supr.warnings = TRUE, omega.freq.method = "pa",
  omega.conf.int.type = "boot", omega.bay.cov.samp = FALSE,
  prior.samp = FALSE, item.dropped = FALSE,
  alpha.int.analytic = FALSE, bayes = TRUE, freq = TRUE,
  para.boot = FALSE, boot.interval.type = "basic", jags = FALSE)
}
\description{
function to calculate all the internal consitency estimates
takes as input both datasets, either matrix or frame, and covariance matrices
so far input of a covariance matrix is not supported
}
