% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omegaFit.R
\name{seco_fit}
\alias{seco_fit}
\title{graphical posterior predictive check for the higher-order factor model,}
\usage{
seco_fit(x, data)
}
\arguments{
\item{x}{A bomegas output object (list)}

\item{data}{A matrix or data.frame containing the data set that produced x}
}
\description{
gives posterior predictive check for the higher-factor model:
comparison between posterior sample of model implied covariance matrices
and sample covariance matrix. Gray bars should enclose the black dots for good fit.
}
\examples{
seco_fit(bomegas(upps, n.factors = 5, n.chains = 2, n.iter = 100,
n.burnin = 50, missing = "listwise"), upps)

}
