% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{secoFit-deprecated}
\alias{secoFit-deprecated}
\title{model fit for the second-order factor model,}
\usage{
secoFit(x, data, ppc = TRUE, cutoff = .08, ci = .90)
}
\arguments{
\item{x}{A bomegas output object (list)}

\item{data}{A matrix or data.frame containing the data set that produced x}

\item{ppc}{A logical indicating if the PPC should be printed or not,
the default is TRUE}

\item{cutoff}{A value to compare the posterior sample of RMSEAs against.
The result will contain the probability that the RMSEA is smaller than the
cutoff value}

\item{ci}{A value between 0 and 1 indicating the credible interval for the RMSEA}
}
\description{
Fit indices and posterior predictive check for the higher-factor model:
comparison between posterior sample of model implied covariance matrices
and sample covariance matrix. Gray bars should enclose the black dots for good fit.
Also prints fit indices, LR (likelihood-ratio), RMSEA, SRMR.
The RMSEA is from Garnier-Villareal & Jorgensen (2020)
}
\references{
{
\insertRef{Garnier-Villarreal2020AdaptingFitIndices}{Bayesrel}
}
}
\seealso{
\code{\link{Bayesrel-deprecated}}
}
\keyword{internal}
