\name{upps}
\alias{upps}
\docType{data}
\title{
20-item questionnaire data from Lozano et al. (2018)
}
\description{
A dataset consisting of 455 participants who filled out the 20-item short form of
the UPPS-P, and impulsivity scale, rating from 0 to 4 on a Likert scale.
The scale has five subscales measured by four items each:
negative urgency (columns 1-4), perserverance (columns 5-8), premeditation (columns 9-12),
sensation seeking (columns 13-16), positive urgency (columns 17-20).
The data contain 13 missing values.}

\usage{upps}
\format{
  The format is a 20-column datamatrix containing 455 observations
}
\source{
article}
\references{
Lozano, Ó. M., Díaz-Batanero, C., Rojas,  A.  J., Pilatti, A., & Fernández-Calderón, F. (2018).
Concordance between the original and short version of the Impulsive Behaviour Scale UPPS-P using an IRT model.
PLOS ONE,13(3), 1–15. https://doi.org/10.1371/journal.pone.0194390
}

\keyword{datasets}
