% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_apply_functions.R
\name{applyalpha}
\alias{applyalpha}
\title{the basic functions for calculating and bootstrapping the internal consistency estimates}
\usage{
applyalpha(M)
}
\description{
the basic functions for calculating and bootstrapping the internal consistency estimates
}
