% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_reliability.R
\name{brel}
\alias{brel}
\title{function to calculate all the internal consitency estimates}
\usage{
brel(raw.data, boot.n = 200, interval = 0.95,
  boot.interval.type = "basic", jags = FALSE, n.iter = 2000,
  n.burnin = 50, freq = TRUE, estimates = c("alpha", "l2", "l4", "l6",
  "glb", "omega"), supr.warnings = TRUE, omega.freq.method = "pa",
  omega.conf.type = "boot", omega.cov.samp = TRUE, returnSamples = FALSE,
  prior.samp = FALSE, if.item.dropped = FALSE)
}
\description{
function to calculate all the internal consitency estimates
}
