% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_sampler.R
\name{covSamp}
\alias{covSamp}
\title{this function uses gibbs sampling to estimate the posterior distribution
of a sample's covariance matrix
sources: https://en.wikipedia.org/wiki/Normal-inverse-Wishart_distribution,
Murphy, K. P. (2007).Conjugate bayesian analysis of the gaussian distribution(Tech. Rep.). University of British Columbia.
Gelman, A., Stern, H. S., Carlin, J. B., Dunson, D. B., Vehtari, A., & Rubin, D. B. (2013). Bayesian data analysis. Chapman and Hall/CRC.}
\usage{
covSamp(data, n.iter, n.burnin)
}
\description{
this function uses gibbs sampling to estimate the posterior distribution
of a sample's covariance matrix
sources: https://en.wikipedia.org/wiki/Normal-inverse-Wishart_distribution,
Murphy, K. P. (2007).Conjugate bayesian analysis of the gaussian distribution(Tech. Rep.). University of British Columbia.
Gelman, A., Stern, H. S., Carlin, J. B., Dunson, D. B., Vehtari, A., & Rubin, D. B. (2013). Bayesian data analysis. Chapman and Hall/CRC.
}
