% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_split_half.R
\name{MaxSplitHalfHad12}
\alias{MaxSplitHalfHad12}
\title{max split half by using hadamard matrices as starting points
this is the solution that is supposed to be faster.
yet when the covariance matrix has lots of negative entries, it doesnt produce a solution}
\usage{
MaxSplitHalfHad12(M)
}
\description{
max split half by using hadamard matrices as starting points
this is the solution that is supposed to be faster.
yet when the covariance matrix has lots of negative entries, it doesnt produce a solution
}
