% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_gibbs_sampler.R
\name{omegaSampler}
\alias{omegaSampler}
\title{this function uses gibbs sampling to estimate the loadings and error variances
of a cfa one factor model
it returns the posterior distribution sample of omegas calculated from those parameters}
\usage{
omegaSampler(data, n.iter, n.burnin)
}
\description{
this function uses gibbs sampling to estimate the loadings and error variances
of a cfa one factor model
it returns the posterior distribution sample of omegas calculated from those parameters
}
