% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_strel_id}
\alias{plot_strel_id}
\title{plots posterior distributions of chosen estimate and the item-dropped cases in one plot}
\usage{
plot_strel_id(x, estimate, distance = NULL)
}
\arguments{
\item{x}{A strel output object (list)}

\item{estimate}{A character string indicating what estimate to plot from the strel output object}

\item{distance}{A string indicating what distance measure is to be used for ordering, if NULL then no ordering,
otherwise ca be either "mean" for the differences between the point estimates;
"ks" for the Kolmogorov Smirnov distance;
or "kl" for the Kublack Leibler divergence}
}
\description{
gives posterior densities of original dataset together with the the posteriors of datasets with
items deleted. Can be ordered for the change item deleting brings about. The method for ordering can be chosen;
the items are ordered in from top to bottom from most change to least change in the posterior
}
\examples{
plot_strel_id(strel(asrm, "lambda2", freq = FALSE, item.dropped = TRUE, n.chains = 2), "lambda2")

}
