% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plotifitem_all}
\alias{plotifitem_all}
\title{plots posterior distributions of chosen estimate and the item-dropped cases in one plot}
\usage{
plotifitem_all(x, estimate, ordering = FALSE)
}
\description{
plots posterior distributions of chosen estimate and the item-dropped cases in one plot
}
