% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plotIfItem_one}
\alias{plotIfItem_one}
\title{plotting function of the posterior for the if-item-dropped statistics}
\usage{
plotIfItem_one(x, estimate, item.pos, criteria = TRUE,
  blackwhite = FALSE, top.align = FALSE, greek = FALSE,
  cuts = c(0.7, 0.8))
}
\description{
plotting function of the posterior for the if-item-dropped statistics
}
