% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_Rel.R
\name{probRel}
\alias{probRel}
\title{this function takes a mcmc posterior sample of any of the internal consistency estimates
and calculates any given probability of the estimate being bigger
or smaller than an arbitrary value}
\usage{
probRel(x, low.bound, estimate)
}
\description{
this function takes a mcmc posterior sample of any of the internal consistency estimates
and calculates any given probability of the estimate being bigger
or smaller than an arbitrary value
}
