% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_reliability.R
\name{rel}
\alias{rel}
\title{calculate all the internal consistency estimates
takes as input both datasets, either matrix or frame, and covariance matrices
so far input of a covariance matrix is not supported}
\usage{
rel(x, estimates = c("alpha", "lambda2", "lambda6", "glb", "omega"),
  interval = 0.95, omega.freq.method = "pa", omega.fit = FALSE,
  alpha.int.analytic = FALSE, bayes = TRUE, freq = TRUE,
  cor.mat = TRUE, para.boot = FALSE, prior.samp = FALSE,
  item.dropped = FALSE, n.iter = 2000, n.burnin = 50,
  boot.n = 1000, supr.warnings = TRUE)
}
\description{
calculate all the internal consistency estimates
takes as input both datasets, either matrix or frame, and covariance matrices
so far input of a covariance matrix is not supported
}
